refactor: 이미지 업로드 관련 함수들 커스텀훅으로 분리
